# swagger/swagger.json
{"swagger":"2.0","info":{"title":"You and local government","description":"A platform for interacting with you and local government","version":""},"host":"localhost:3000","schemes":["http"],"consumes":["application/json","application/xml","application/gob","application/x-gob"],"produces":["application/json"],"paths":{"/admin/councillor":{"post":{"tags":["admin"],"summary":"createCouncillor admin","description":"admin api to add a councillor\n\nRequired security scopes:\n  * `admin:access`","operationId":"admin#createCouncillor","responses":{"201":{"description":"Created"},"401":{"description":"Unauthorized"}},"schemes":["http"],"security":[{"jwt":["admin:access"]}]}},"/swagger-ui/{filepath}":{"get":{"summary":"Download swagger-ui/","operationId":"swagger#/swagger-ui/*filepath","parameters":[{"name":"filepath","in":"path","description":"Relative file path","required":true,"type":"string"}],"responses":{"200":{"description":"File downloaded","schema":{"type":"file"}},"404":{"description":"File not found","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}},"/swagger.json":{"get":{"summary":"Download swagger/swagger.json","operationId":"swagger#/swagger.json","responses":{"200":{"description":"File downloaded","schema":{"type":"file"}}},"schemes":["http"]}},"/user":{"get":{"tags":["user"],"summary":"list user","description":"get a list user\n\nRequired security scopes:\n  * `api:access`","operationId":"user#list","responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/GoaLocalUserCollection"}},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"}},"schemes":["http"],"security":[{"jwt":["api:access"]}]},"post":{"tags":["user"],"summary":"create user","description":"admin api to add a user\n\nRequired security scopes:\n  * `admin:access`","operationId":"user#create","parameters":[{"name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/User"}}],"responses":{"201":{"description":"Created"},"404":{"description":"Not Found"}},"schemes":["http"],"security":[{"jwt":["admin:access"]}]}},"/user/login":{"post":{"tags":["user"],"summary":"login user","description":"login user","operationId":"user#login","parameters":[{"name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/Login"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/GoaLocalUser"}}},"schemes":["http"]}},"/user/signup":{"post":{"tags":["user"],"summary":"signup user","description":"Signup a user","operationId":"user#signup","parameters":[{"name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/User"}}],"responses":{"201":{"description":"Created"},"404":{"description":"Not Found"}},"schemes":["http"]}},"/user/{id}":{"get":{"tags":["user"],"summary":"read user","description":"get a user\n\nRequired security scopes:\n  * `api:access`","operationId":"user#read","parameters":[{"name":"id","in":"path","description":"user ID","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/GoaLocalUser"}},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"}},"schemes":["http"],"security":[{"jwt":["api:access"]}]},"post":{"tags":["user"],"summary":"update user","description":"update a user\n\nRequired security scopes:\n  * `api:access`","operationId":"user#update","parameters":[{"name":"id","in":"path","description":"user ID","required":true,"type":"string"},{"name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/UpdateUser"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/GoaLocalUser"}},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"}},"schemes":["http"],"security":[{"jwt":["api:access"]}]},"delete":{"tags":["user"],"summary":"delete user","description":"delete a user\n\nRequired security scopes:\n  * `api:access`","operationId":"user#delete","parameters":[{"name":"id","in":"path","description":"user ID","required":true,"type":"string"}],"responses":{"202":{"description":"Accepted"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"}},"schemes":["http"],"security":[{"jwt":["api:access"]}]}}},"definitions":{"GoaLocalUser":{"title":"Mediatype identifier: application/vnd.goa.local.user+json; view=default","type":"object","properties":{"area":{"type":"string","description":"The area of the users local council","example":"Velit quia commodi dolorem."},"email":{"type":"string","description":"The email of the user","example":"In harum dolores consequatur et."},"firstName":{"type":"string","description":"Name of the user","example":"Ipsum optio magnam eos."},"href":{"type":"string","description":"API href for making requests on the bottle","example":"Et non similique."},"id":{"type":"string","description":"Unique bottle ID","example":"Eos earum vitae magnam sunt."},"secondName":{"type":"string","description":"Name of the user","example":"Recusandae aut."},"signupType":{"type":"string","description":"the signupType of user google local","example":"Id recusandae qui mollitia blanditiis corrupti eum."},"type":{"type":"string","description":"the type of user admin local councillor","default":"local","example":"Aut corporis et."}},"description":"A User of locals (default view)","example":{"area":"Velit quia commodi dolorem.","email":"In harum dolores consequatur et.","firstName":"Ipsum optio magnam eos.","href":"Et non similique.","id":"Eos earum vitae magnam sunt.","secondName":"Recusandae aut.","signupType":"Id recusandae qui mollitia blanditiis corrupti eum.","type":"Aut corporis et."},"required":["firstName","secondName","email"]},"GoaLocalUserCollection":{"title":"Mediatype identifier: application/vnd.goa.local.user+json; type=collection","type":"array","items":{"$ref":"#/definitions/GoaLocalUser"},"description":"GoaLocalUserCollection is the media type for an array of GoaLocalUser (default view)","example":[{"area":"Velit quia commodi dolorem.","email":"In harum dolores consequatur et.","firstName":"Ipsum optio magnam eos.","href":"Et non similique.","id":"Eos earum vitae magnam sunt.","secondName":"Recusandae aut.","signupType":"Id recusandae qui mollitia blanditiis corrupti eum.","type":"Aut corporis et."},{"area":"Velit quia commodi dolorem.","email":"In harum dolores consequatur et.","firstName":"Ipsum optio magnam eos.","href":"Et non similique.","id":"Eos earum vitae magnam sunt.","secondName":"Recusandae aut.","signupType":"Id recusandae qui mollitia blanditiis corrupti eum.","type":"Aut corporis et."},{"area":"Velit quia commodi dolorem.","email":"In harum dolores consequatur et.","firstName":"Ipsum optio magnam eos.","href":"Et non similique.","id":"Eos earum vitae magnam sunt.","secondName":"Recusandae aut.","signupType":"Id recusandae qui mollitia blanditiis corrupti eum.","type":"Aut corporis et."}]},"Location":{"title":"Location","type":"object","properties":{"Lat":{"type":"number","example":0.19498104724300513,"format":"double"},"Lon":{"type":"number","example":0.62452386199952,"format":"double"}},"example":{"Lat":0.19498104724300513,"Lon":0.62452386199952}},"Login":{"title":"Login","type":"object","properties":{"email":{"type":"string","example":"Sed sapiente nemo rerum eum."},"signupType":{"type":"string","example":"Qui dolores dolores incidunt labore."},"token":{"type":"string","example":"Enim nesciunt minima ab."}},"example":{"email":"Sed sapiente nemo rerum eum.","signupType":"Qui dolores dolores incidunt labore.","token":"Enim nesciunt minima ab."},"required":["email","token","signupType"]},"UpdateUser":{"title":"UpdateUser","type":"object","properties":{"area":{"type":"string","description":"The area of the users local council","default":"","example":"Qui velit natus officia reiciendis aut."},"email":{"type":"string","description":"The email of the user","example":"Recusandae perferendis et optio id expedita."},"firstName":{"type":"string","description":"Name of the user","example":"Aliquam exercitationem earum provident qui."},"id":{"type":"string","description":"Unique user ID","example":"Pariatur ut."},"image":{"type":"string","description":"an image url for the user","default":"","example":"Molestiae voluptatem quis voluptates."},"secondName":{"type":"string","description":"Name of the user","example":"Eaque quia ut ut."}},"example":{"area":"Qui velit natus officia reiciendis aut.","email":"Recusandae perferendis et optio id expedita.","firstName":"Aliquam exercitationem earum provident qui.","id":"Pariatur ut.","image":"Molestiae voluptatem quis voluptates.","secondName":"Eaque quia ut ut."},"required":["firstName","secondName","email","id","area"]},"User":{"title":"User","type":"object","properties":{"area":{"type":"string","description":"The area of the users local council","default":"","example":"Cum ea quod accusamus aliquam."},"email":{"type":"string","description":"The email of the user","example":"Voluptatem officia eum natus aut."},"firstName":{"type":"string","description":"Name of the user","example":"Omnis facilis ratione odio."},"id":{"type":"string","description":"Unique user ID","example":"Culpa doloribus."},"image":{"type":"string","description":"an image url for the user","default":"","example":"Earum inventore optio alias rerum debitis ut."},"location":{"$ref":"#/definitions/Location"},"secondName":{"type":"string","description":"Name of the user","example":"Sapiente nihil quod architecto harum dolorum."},"signupType":{"type":"string","description":"the signupType of user google local","default":"","example":"Earum voluptas."},"token":{"type":"string","description":"This can be an oauth token or a password","example":"Ipsa mollitia error impedit minima."},"type":{"type":"string","description":"the type of user admin local councillor","default":"local","example":"Ea ex."}},"example":{"area":"Cum ea quod accusamus aliquam.","email":"Voluptatem officia eum natus aut.","firstName":"Omnis facilis ratione odio.","id":"Culpa doloribus.","image":"Earum inventore optio alias rerum debitis ut.","location":{"Lat":0.19498104724300513,"Lon":0.62452386199952},"secondName":"Sapiente nihil quod architecto harum dolorum.","signupType":"Earum voluptas.","token":"Ipsa mollitia error impedit minima.","type":"Ea ex."},"required":["firstName","secondName","email","token"]},"error":{"title":"Mediatype identifier: application/vnd.goa.error","type":"object","properties":{"code":{"type":"string","description":"an application-specific error code, expressed as a string value.","example":"invalid_value"},"detail":{"type":"string","description":"a human-readable explanation specific to this occurrence of the problem.","example":"Value of ID must be an integer"},"id":{"type":"string","description":"a unique identifier for this particular occurrence of the problem.","example":"3F1FKVRR"},"meta":{"type":"object","description":"a meta object containing non-standard meta-information about the error.","example":{"timestamp":1458609066},"additionalProperties":true},"status":{"type":"string","description":"the HTTP status code applicable to this problem, expressed as a string value.","example":"400"}},"description":"Error response media type (default view)","example":{"code":"invalid_value","detail":"Value of ID must be an integer","id":"3F1FKVRR","meta":{"timestamp":1458609066},"status":"400"}}},"responses":{"Accepted":{"description":"Accepted"},"Created":{"description":"Created"},"NotFound":{"description":"Not Found"},"OK":{"description":"OK","schema":{"$ref":"#/definitions/GoaLocalUser"}},"Unauthorized":{"description":"Unauthorized"}},"securityDefinitions":{"jwt":{"type":"apiKey","description":"\n\n**Security Scopes**:\n  * `api:access`: API access","name":"Authorization","in":"header"}}}